Move `btc` and `stx` ledger functions to separate files inside `ledger` folder
